# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Course Management Platform API",
    "version": "1.0.0",
    "description": "API documentation for the Course Management Platform backend"
  },
  "servers": [
    {
      "url": "http://localhost:5000"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
    "/activity-logs": {
      "post": {
        "summary": "Submit a new activity log",
        "tags": [
          "ActivityTracker"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "allocationId": {
                    "type": "integer"
                  },
                  "week": {
                    "type": "integer"
                  },
                  "gradingCompleted": {
                    "type": "boolean"
                  },
                  "attendanceTaken": {
                    "type": "boolean"
                  },
                  "deadlineCommunicated": {
                    "type": "boolean"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Activity log submitted"
          }
        }
      },
      "get": {
        "summary": "Get all activity logs (manager only)",
        "tags": [
          "ActivityTracker"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "facilitatorId",
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "courseId",
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "week",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of all activity logs"
          }
        }
      }
    },
    "/activity-logs/{id}": {
      "put": {
        "summary": "Update an existing activity log",
        "tags": [
          "ActivityTracker"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActivityLog"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Activity log updated"
          }
        }
      },
      "delete": {
        "summary": "Delete a specific activity log",
        "tags": [
          "ActivityTracker"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Log deleted"
          }
        }
      }
    },
    "/activity-logs/my-logs": {
      "get": {
        "summary": "Get own activity logs",
        "tags": [
          "ActivityTracker"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Returns logs of the logged-in facilitator"
          }
        }
      }
    },
    "/allocations": {
      "get": {
        "summary": "Get all allocations",
        "tags": [
          "Allocation"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of allocations"
          }
        }
      },
      "post": {
        "summary": "Create a new allocation",
        "tags": [
          "Allocation"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "201": {
            "description": "Allocation created"
          }
        }
      }
    },
    "/allocations/{id}": {
      "get": {
        "summary": "Get allocation by ID",
        "tags": [
          "Allocation"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Allocation data"
          }
        }
      },
      "put": {
        "summary": "Update an allocation",
        "tags": [
          "Allocation"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Allocation updated"
          }
        }
      },
      "delete": {
        "summary": "Delete an allocation",
        "tags": [
          "Allocation"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Allocation deleted"
          }
        }
      }
    },
    "/allocations/facilitator/dashboard": {
      "get": {
        "summary": "Facilitator dashboard data",
        "tags": [
          "Allocation"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Dashboard info for facilitator"
          }
        }
      }
    },
    "/allocations/manager/dashboard": {
      "get": {
        "summary": "Manager dashboard data",
        "tags": [
          "Allocation"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Dashboard info for manager"
          }
        }
      }
    },
...
    "/classes": {
      "get": {
        "summary": "Get all classes",
        "tags": [
          "Classes"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of classes"
          }
        }
      },
      "post": {
        "summary": "Create a class",
        "tags": [
          "Classes"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Class"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Class created"
          }
        }
      }
    },
    "/classes/{id}": {
      "get": {
        "summary": "Get a class by ID",
        "tags": [
          "Classes"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Class object"
          }
        }
      },
      "put": {
        "summary": "Update a class",
        "tags": [
          "Classes"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Class"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Class updated"
          }
        }
      },
      "delete": {
        "summary": "Delete a class",
        "tags": [
          "Classes"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Class deleted"
          }
        }
      }
    },
...
  },
...
}